fix: allow use of margin notes in paragraphs with drop caps
I was using the Dropcap component to wrap the entire first paragraph. This was so I could wrap the whole thing in a paragraph tag to overcome the MDX issue where they
are missed out. But it's better, for now, to
just wrap a <p> in the .mdx file than to
struggle with a solution for the CSS counters